Fix bug and refactor entity

+++ src/modules/comic/comic.service.ts
@@ -182,12 +182,9 @@ export class ComicService {
   async delete(comicId: number) {
     const comic = await this.getComicById(comicId);
 
-    await this.chapterService.deleteAllChapterOfComic(comicId);
-    // await Promise.all([
-    //   this.userService.deleteComicFromEvaluate(comicId),
-    //   this.userService.deleteComicFromFollow(comicId),
-    //   this.userService.deleteComicFromLike(comicId),
-    // ]);
+    if (!comic) {
+      throw new HttpException(`Cannot found comic with id [${comicId}]`, HttpStatus.NOT_FOUND);
+    }
 
     return await this.comicRepository.delete(comicId);
   }
@@ -219,16 +216,27 @@ export class ComicService {
   }
 
   async getComicBySlug(slugComic: string) {
-    return await this.comicRepository
-      .createQueryBuilder('comic')
-      .where('comic.slug = :slug', { slug: slugComic })
-      .getOne();
+    return await this.comicRepository.findOne({
+      where: {
+        slug: slugComic,
+      },
+      order: {
+        chapters: {
+          order: 'DESC',
+        },
+      },
+    });
   }
 
   async getComicById(comicId: number): Promise<Comic> {
     const comic = await this.comicRepository.findOne({
       where: {
         id: comicId,
+      },
+      order: {
+        chapters: {
+          order: 'DESC',
+        },
       },
     });
 
@@ -418,11 +426,7 @@ export class ComicService {
       throw new HttpException('Truyện không tồn tại!', HttpStatus.NOT_FOUND);
     }
 
-    return this.commentService.createNewComment({
-      userId,
-      comicId,
-      content,
-    });
+    return this.commentService.createNewComment(userId, comic, content);
   }
 
   async addAnswerToCommentOfComic(
@@ -437,11 +441,11 @@ export class ComicService {
       throw new HttpException('Bình luận không tồn tại!', HttpStatus.NOT_FOUND);
     }
 
-    if (comicId !== comment.comicId) {
+    if (comicId !== comment.comic.id) {
       throw new HttpException('Bình luận không thuộc truyện tranh này!', HttpStatus.BAD_REQUEST);
     }
 
-    const comic = await this.getComicById(comment.comicId);
+    const comic = await this.getComicById(comment.comic.id);
 
     if (!comic) {
       throw new HttpException('Truyện không tồn tại!', HttpStatus.NOT_FOUND);
@@ -459,6 +463,10 @@ export class ComicService {
     const comics = await this.comicRepository.find({
       where: {
         creator: userId,
+      },
+      order: {
+        updatedAt: 'ASC',
+        id: 'ASC',
       },
     });
